Fix half/float resolution
Summary:
Instead of the existing and somewhere unpredictable heuristics,
all outputs which are ambiguously half/float are marked as such
and resolved based on the first known half/float output type,
or in the absense, float.

The type inference happens as part of the "canonicalize" step,
alongside the implicit casting.

A separate pass resolves the half or float, after canonicalization.
Prior to this step, ambiguous types will be marked Float16/Float32
in the trace.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T26434

Differential Revision: https://phabricator.sourcevertex.net/D33907