Warn on non-tensor buffers/params
Summary:
After trying various ops, it appears that there are no cases in which
a buffer would be anything other than a TensorType once traced, even
though the source code would appear to permit NoneType. For this
reason, it is safer to have a printed warning to assist debugging
in the event that this situation does occur.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, hatemh

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, hatemh

Maniphest Tasks: T26350

Differential Revision: https://phabricator.sourcevertex.net/D34375