Keep (host side) constants that are simply returned within poptorch
Summary:
This aims to prevent a Popart exception such as
"Anchor tensor `...' not in Ir Tensors"
in cases where the tensor is not used by a Popart op. However, this
will not work until constexpr evaluation is implemented because
the at::detach leads do a popart::detach which is not removed.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, stephenm

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, stephenm

Subscribers: stephenm

Maniphest Tasks: T27798

Differential Revision: https://phabricator.sourcevertex.net/D35121