popart_compiler links popart-only; in custom ops, set policy to check `Protobuf_ROOT` before finding popart
Summary:
This should work without any modification in Popart.
Once this is landed, Popart can switch to having both popart-only and popart components; the latter of which publicly links onnx.
Thanks to this diff, we will be able to land the Popart diff without breaking Poptorch.

Test Plan: Worked on a standalone build locally. Would be good if others confirm this.

Reviewers: anthonyb, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, vladimirm

Reviewed By: anthonyb, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch

Maniphest Tasks: T29285

Differential Revision: https://phabricator.sourcevertex.net/D36623