Add support for multiconv
Summary:
Support targeting the popart implementation of MultiConv from poptorch.  This solution uses a context manager to mark the boundaries for which convolutions to combine into a multiconv.  Then swap out any conv for a multi_conv_part that is eventually used to build up the entire multiconv in the compiler.

Note this now depends on D36748 landing in popart first.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T28354

Differential Revision: https://phabricator.sourcevertex.net/D36103