Add support for interpolate with "nearest" mode
Summary: `torch.nn.functional.interpolate` shares the same handler with `upsample` but omits the optional scale inputs.  This diff changes the upsample handler to compute the scales using ratio of the `output_size` to the `input_size` which works for both `interpolate` and `upsample` ops.

Test Plan:
* Enabled additional tests
* Tested with Octave Convolution example https://github.com/braincreators/octconv

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Maniphest Tasks: T31607

Differential Revision: https://phabricator.sourcevertex.net/D37736