Create shallow copy of model
Summary:
Currently, because the way the training and inference wrapper are set up, all changes made to the local variable `_model` and `_user_model` in the `PoplarExecutor` end up being reflected to the user model that is wrapped. While it is desirable to keep the weights in sync, adding to it or modyifying its class to PoptorchModel are unwanted by-products.

This diff fixes this by making a shallow copy of the model, keeping a reference to whatever is in the user model, which will allow for keeping the weights synced, while disconnecting it from changes made in the `PoplarExecutor`.

As a consequence it will now always be possible to use `torch.save` and `torch.load` with the user model, even as the wrapped model may not be pickle-friendly.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T31288

Differential Revision: https://phabricator.sourcevertex.net/D37424