Support torch.min/max with dim and keepdim inputs
Summary: This patch adds support for torch.min and torch.max with dim and keepdim inputs.  This overload returns a tuple with the values as well as their indices (argmax/argmin).  This implementation uses the popart topk op with k == 1 since that operator with also return both the values and indices.

Test Plan: Added tests to math_ops_test.py

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Maniphest Tasks: T34502

Differential Revision: https://phabricator.sourcevertex.net/D40861