Fix: separate compilation from device initialisation on distributed system
Summary:
All processes must be initialised at the same time as they need to communicate with each other.
At the moment the first process will grab the lock and compile the executable but then it will proceed directly to initialise its device but as the other processes are still waiting for the lock to be released it will dead lock.
Unfortunately there is currently no way to separate compilation from device initialisation in Popart so as a workaround the first process to grab the lock will just do a dummy call to compileAndExport to populate the cache.
Then all the processes will call compile() at the same time but should all get a cache hit and can then proceed with their respective device initialisation.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, hatemh, nicholasw

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, hatemh, nicholasw

Maniphest Tasks: T35382

Differential Revision: https://phabricator.sourcevertex.net/D41691