Support inplace modifications to inputs
Summary:
Support inplace ops by updating the input variable which is modified
in place with the output variable: this support is emulated and
PopART is free to actually carry out the op in place or not depending
on what is optimal according the heuristics. Any inplace ops which
would not change and input may not be inplace as this is
inconsequential to the result.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb, jamiep

Maniphest Tasks: T32386

Differential Revision: https://phabricator.sourcevertex.net/D43883