Refactor ModelWithWeights to take arbitrary num inputs
Summary:
For ops with >1 input, only the first input will be passed through the linear layer. The output of the lin layer along with the rest of the original inputs is then passed to the op.

Added tests for a couple of ops that were missed in previous diffs

  - TripletMarginLoss
  - CTCLoss

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T38220

Differential Revision: https://phabricator.sourcevertex.net/D46179