Allow a constant to be both PopART and host side
Summary:
Allow a constant which is both returned host side (possibly in a
tuple/list), and used in PopART. Previously, this would lead to an
error. This works by splitting the constant into two, a popart
constant and host side constant, and altering any users to use the
correct constant.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Subscribers: ahmads

Maniphest Tasks: T41774

Differential Revision: https://phabricator.sourcevertex.net/D47828