Use numerically stable softplus
Summary: This changes the softplus handler to use the recently added implementation in PopART which is more numerically stable than evaluating the softplus directly using the transcendental functions.

Test Plan: I checked that the added test function fails with the lastest SDK with an overflow for `torch.nn.Softplus`.  Also opened T44221 to investigate other activation functions that seem to fail with FP exceptions on but give the correct answer when FP exceptions are turned off.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, mihailp

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, mihailp

Maniphest Tasks: T44164

Differential Revision: https://phabricator.sourcevertex.net/D49808