Allow PopTorch to build and run against PyTorch 1.10/nightlies.
Summary: It's useful to see what changes in future versions of torch so we should maintain our codebase to be compatable with both when it is possible. This should allow us to test previous revisions of torch in the future and will make it easier to move to 1.10 when we do.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, mihailp

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, mihailp

Maniphest Tasks: T24504

Differential Revision: https://phabricator.sourcevertex.net/D50238