Fix spurious error messages when running Async Dataloaders in Jupyter
Summary:
This is a fix for T48028 and its parent T46957 (**Customer**: DeepSenseAI, ZD Priority: Normal)

When a `poptorch.DataLoader` is run in `Async` mode in a notebook it causes an error to be printed to the console when the kernel is shut down, which is not a fantastic user experience.

A test is added reproducing the issue and a fix is applied.

A couple of references to Pytorch may be useful to consider:
- the fix in this diff is similar to the structure of [the multiprocessing worker used by pytorch](https://github.com/pytorch/pytorch/blob/aa7da7b09c4a3f972ede5fd8ad0cbc8c13498a00/torch/utils/data/_utils/worker.py#L202) (wrapping the whole worker loop in a `try ... except KeyboardInterrupt ... ` block).
- the worker process shutdown seems to be an area where they tread carefully in PyTorch:
https://github.com/pytorch/pytorch/blob/aa7da7b09c4a3f972ede5fd8ad0cbc8c13498a00/torch/utils/data/dataloader.py#L570

Test Plan: A pure python test which reproduces the jupyter error has been added, and a fixed is added. Tests were run on a POD before and after and no new failures were observed.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T48028

Differential Revision: https://phabricator.sourcevertex.net/D53685