Add option broadcastBuffers to Poptorch and allow the use of independant buffers in replicas
Summary:
An option to allow replicas to use buffers (without sync).

Here's the plan:

- add an option for broadcast buffers whose default is True (poptorch.Options.Distributed.broadcastBuffers(True));
- trigger an error when all the following are true:
        - there's a buffer in the graph
        - the buffer is modified in place
        - replication is enabled
        - when the option is set to True (i.e. error by default, for now)
- once we support broadcast buffers, remove the error message;

By doing so, we have the same default as PyTorch, and we ensure we will not break any code in the future. Meaning, as soon as now, the customer will have to **explicitly** set the option to False, so they are certain this will not break in the future.

I see in the code that buffers are being seen as "parameters". But in PyTorch's terms, this is not correct. In the inlined graph coming from the tracer, a buffer is a non-constant tensor with requires_grad=0 while a parameter has requires_grad=1. This should be enough to differentiate buffers from parameters, and they should not be confused (parameters are not broadcast for example).

For T47236

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb, stephenm, nicholasw, alexandrec, jamiep

Maniphest Tasks: T47236

Differential Revision: https://phabricator.sourcevertex.net/D54747