Z3390 - [PyTorch] - BUG of poptorch slice after torch.chunk
Summary:
Adding the shape to the output of the torch chunk canonicalisation as the output dimensions are empty.

When the results of a torch.cat is fed into a torch.chunk the canonicalisation step for torch.chunk does not give its outputs any size and therefore later functions do not get canonicalised correctly, for example a later slice returns an identity.

Test Plan: Ran Poptorch tests locally

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, nicholasw, alexandrec, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb, stephenm

Maniphest Tasks: T48271

Differential Revision: https://phabricator.sourcevertex.net/D55089