BuildEnv - Adding general exception to support CentOS
Summary:
T48801 covers fixing the Out of Box tests for CentOs machines - which currently have an error in the create_buildenv script. This diff adds general exception support, which allows the CentOS tests to work

Spoke to @anthonyb to check this was fine to add.

Test Plan: Tested on a local Jenkins scratch space - this allowed CentOS tests to pass

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T48801

Differential Revision: https://phabricator.sourcevertex.net/D55413