Reduce memory footprint of torch.roll
Summary:
In T51234 users are encountering OOMs when using torch roll, but a
naive python implementation of it removes that OOM.

Inspecting the code showed that the input tensor to torch.roll was
being duplicated before being sliced. I suggest we slice first and
concatenate the chunks after.

To the reviewers: Do you want me to look at capturing this difference
in memory behaviour in a test?

Test Plan: Locally, now hoping for CI's approval.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads, nejci

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Maniphest Tasks: T51234

Differential Revision: https://phabricator.sourcevertex.net/D57200