Match the roll API on IPU and CPU
Summary:
This diff exactly matches the roll API for torch on CPU and
IPU. This is done by supporting negative axes values in the
same range as pytorch.
the range [-len, len-1].

To test this functionality the op_harness is extended to
support matching exceptions as well as tensors.

The broader goal is to provide a testing function which provided
with an op and arbitrary inputs for that op will reliably check
if the torch and poptorch APIs remain equivalent.

Because we use jit trace, once we support all the same input
ranges as pytorch we get the exact same exceptions
for free as the argument checking lives in the python part
of the upstream package which gets executed.

Bonus change:
Replace slices with a split in the roll API

T51234

Test Plan: Locally all tensor_ops_test pass

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Maniphest Tasks: T51234

Differential Revision: https://phabricator.sourcevertex.net/D57614