Convert halves to floats before calling CPU op implementation in JIT dispatch
Summary:
We use the CPU implementations of ops to do shape inference in the JIT dispatcher. This causes
a problem with half-typed networks as the CPU implementations often don't handle half-typed
inputs. However, we only care about the shape, so this patch changes any half inputs into floats
at the point of calling the CPU implementation. The outputs will be fixed back to the correct types
by canonicaliseAndFixOutput().

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Maniphest Tasks: T44167

Differential Revision: https://phabricator.sourcevertex.net/D60241