Use output shape instead of input shape in flip handler
Summary: The original fix was using the input shape to determine the offset to convert negative indices to positive indices. However, if the input is the result of another operation, the shape will be lost, so no offset is added. This fix uses the output shape instead, since input shape == output shape for this operation, and the output shape is always available

Test Plan: Works with reproducer provided by Apps

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: jamesbr

Maniphest Tasks: T56015

Differential Revision: https://phabricator.sourcevertex.net/D61546