Adapt constant canonicalisation for dispatch
Summary: I split the pass into two steps, one working on a per-node-basis and it's stateless (called in `createAtenTarget`) and one that needs a full graph (called in `compileWithManualTracing`). I could try to make it a state machine and fit everything into a call at `createAtenTarget` but I think it would add a lot of unnecessary complexity.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Subscribers: ahmads

Maniphest Tasks: T57406

Differential Revision: https://phabricator.sourcevertex.net/D62634