Fix the incorrect naming of is_const to is_empty
Summary:
`is_const` should really be named `is_empty`, and all new tensors created
by `registerEmptyTensor()` should have it set. If an empty tensor is used
to store the output of a node, the flag should be cleared. If an empty
tensor is passed directly to a node as input, that means it's a constant
(following the logic, anything that isn't an input tensor or an output
from a node is a constant).

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Maniphest Tasks: T51159

Differential Revision: https://phabricator.sourcevertex.net/D63538