Add a test for torch clamp for all data types
Summary:
Not skipping checkAndChangeOutputTypes for ListUnpack appears to fix the `int64` and not cause any unit test failures.
This appears to fix 1 of 2 problem cases in the original repro.py.
Appears to fix the `int64` case (and unintentionally the `float64` case)

Still need to fix the `float16` case. Is this supported? It doesn't appear in `MANY_TYPES` in `poptorch/tests/type_support_test.py`
I'm struggling to get a `float16` model to compile, so would be nice to chat with someone from the Poptorch team please.

Would welcome any feedback on what cases the code I've deleted was intended to cover?

Test Plan: See `type_support_test.py`

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, nicholasw, erikn, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, nicholasw, anthonyb

Subscribers: anthonyb, nicholasw

Maniphest Tasks: T58807

Differential Revision: https://phabricator.sourcevertex.net/D64166