Exclude zero from range of uniform distribution in exponentialHandler
Summary:
In the current implementation of exponential_() there's a log operation performed on values drawn from a uniform distribution [0, 1], and this means that it's possible to have a log(0), which gives -inf. This change prevents that by sampling from an interval that doesn't include 0.
Fix T58658.

Test Plan: CI + reproducer in task

Reviewers: #pytorch, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ahmads, manueles

Reviewed By: #pytorch, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, ahmads

Subscribers: davidw, ahmads

Maniphest Tasks: T58658

Differential Revision: https://phabricator.sourcevertex.net/D63436