Fix error messages when overlapping input and output
Summary:
When both output and input are overlapped these will share a constant string
node that represents the input (due to an optimization for sharing constants
with the same value). We need to be careful to only delete this node when
no-one else is using it.

This also fixes a potential bug if the user uses a string with this name in
their model

Test Plan: Added tests for having errors when overlapping input and output

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, nicholasw, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb

Maniphest Tasks: T55819

Differential Revision: https://phabricator.sourcevertex.net/D66655