Clean up - part 1
Summary:
Separated compilation from execution a bit better:
- Executable: it's the result of the compilation process (i.e a binary), to generate an Executable we need a graph + a Target
- Executor: it loads an Executable onto a Device, and manages its execution.
- ExecutorWrapper: we can't have poplar types leak in the Torch code so this is a wrapper around the Executor.

Compilation now happens through a function rather than through the Executable class: as explained above, the Executable should only contain the result of the compilation, not anything related to the compilation process (This will make things easier when we start implementing executable caching).

For now we've got PoplarExecutable / PoplarExecutor / PoplarExecutorWrapper (And we'll have Popit variants in the future).

I've tried to clean up a bit the pimpl and tried to make sure the filenames matched the class names.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Subscribers: ahmads

Maniphest Tasks: T57253

Differential Revision: https://phabricator.sourcevertex.net/D68135