Make OnDemand devices attach to any matching device.
Summary:
If we rely on PopART's implementation of OnDemand it will only try to connect to the first IPU matching the config which creates contention on IPU 0 when several tests only need 1 IPU for example.
This patch implements a workaround: if the connection type is OnDemand and we're not attached to any device, then when we do need a device create a new DeviceManager, switch to Always, and request a new device.
We then swap the devices in the live session, and resume execution from there.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Subscribers: davidw

Maniphest Tasks: T65094

Differential Revision: https://phabricator.sourcevertex.net/D70330