Resolve broken test, add `test_linear`
Summary:
This test uses a LogChecker to find elements in the graph and ensures
they have the correct phase set.

It appears the new dispatcher prefers to reshape in to a MatMul and reshape
back which trips up this test. In addition, the last MatMul is replaced
with a Gemm.

This revision also includes a new test for `torch.functional.linear`
based on the inputs given in the other test.

This revision was originally reverted as I'd used `torch.empty` to generate
the weight and bias tensors. They are now `torch.ones`.

Ref T57195

Test Plan: Existing test works, adds new test which should also pass CI.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Maniphest Tasks: T57195

Differential Revision: https://phabricator.sourcevertex.net/D71449