Remove tracker
Summary:
As well as removing the tracker, this patch makes other changes to support this.

Firstly, it moves tensor information that we want to store in the `ValueMapper` into
its own structure, which is referenced by a `shared_ptr` from `IpuTensorImpl`. This
way, we don't need to hang on to the `IpuTensorImpl` itself, so we avoid clashing with
`ExclusivelyOwned`.

Secondly, the patch also changes `ValueMapper` to key its maps with a pointer to the new
structure rather than the `TensorImpl` pointer, which is no longer guaranteed to be unique,
as we're no longer holding on to an `intrusive_ptr`. `JITDispatch` also now delegates its tensor
creation to its wrapped `MLIRDispatch`, so we avoid having two sources of IPU tensor IDs, which
leads to ID collision.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb

Subscribers: anthonyb, davidw

Maniphest Tasks: T57195

Differential Revision: https://phabricator.sourcevertex.net/D70356