Fix `tensor.detach()` when using the dispatcher
Summary:
This intercepts the `detach` op when using the dispatcher just the same as we
already do for the tracer, but with one special case to handle moving
parameters from the CPU. `cpu_tensor.to("xla")` invokes this detach but we
still want Autograd to traverse backwards through these parameters, so we'll
simply shallow copy in this case.

Ref T57195

Requires D70356

Test Plan: Removes the test skip.

Reviewers: anthonyb, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, jackha

Reviewed By: anthonyb, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch

Subscribers: davidw

Maniphest Tasks: T57195

Differential Revision: https://phabricator.sourcevertex.net/D71561