Fix kl_div when target<=0
Summary:
The aten implementation does the following:

```
Tensor kl_div(const Tensor& input, const Tensor& target, int64_t reduction) {
  auto zeros = at::zeros_like(target);
  auto output_pos = target * (at::log(target) - input);
  auto output = at::where(target > 0, output_pos, zeros);
  return apply_loss_reduction(output, reduction);
}
```

We were missing the clamping that happens in the penultimate line of the function. I have added this in the PopART canonicalisation and modified the test so that it also tests the situation when the provided targets are less than or equal to zero.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Subscribers: ahmads

Maniphest Tasks: T66658

Differential Revision: https://phabricator.sourcevertex.net/D73175