PReLU: support PyTorch weight broadcasting
Summary:
PyTorch's implementation of PReLU is not directly compatible with ONNX,
because there is an extra unsqueeze step required to ensure it is
universally broadcastable according to the ONNX spec.

This change removes `prelu` from automatic op generation and adds a
manual handler which appends `(input.ndim - weight.ndim - 1)` singleton
dimensions to the `weight` before passing to ONNX.

Closes T66101

Test Plan: Adds a new PReLU test which demonstrates the original issue.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Maniphest Tasks: T66101

Differential Revision: https://phabricator.sourcevertex.net/D73294