Add `ipu_wrapper` function decorator alternative to `IPUContext`
Summary:
This commit introduces the new `ipu_wrapper` function decorator which serves
as a replacement for `IPUContext` for the MLIR dispatcher, as well as
permitting the conversion of CPU tensors to IPU tensors outside the
graph. They will be added ('promoted') as inputs to the graph when the wrapped
function is called with those arguments. Tensors converted inside will be
added to the graph as constants. Outputs will be promoted when returned.

Also includes:

* Support for casting `host_buffer`s of IPU tensors on the host when the
  dispatcher is not active.

Closes T61527

Test Plan: add a new `ipu_session_test.py`

Reviewers: davidw, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: davidw, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Subscribers: ahmads

Maniphest Tasks: T61527

Differential Revision: https://phabricator.sourcevertex.net/D74123