Target vectorised scatter implementation in PopART from PopTorch
Summary:
This updates the `index_put` handler to target the recently added `None` reduction mode for the PopART `ScatterReduce` operator.  This operator allows a vector of indices to be used to update a multi-dimensional tensor which is more flexible than the ONNX versions of the scatter operator.

Depends on D74229 and tested with the reproducer from @jeromem in T49876. Profiling shows:

* 561 MB reduction in memory usage
* ~100X reduction in the cycle count (from ~2.6e6 to ~1.2e4 cycles)

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, ahmads

Subscribers: jeromem, ahmads

Maniphest Tasks: T49876

Differential Revision: https://phabricator.sourcevertex.net/D74254