Refactored how view operations work and added side effect annotations so the mlir DCE and CSE can be used
Summary:
This commit enables the dead code elimination (DCE) and common subexpression elimination (CSE) passes from mlir. For this to work on every operation we needed to convert all the reference semantics in the graph to value semantics.

For inplace operations this is implemented by replacing the inplace op with the outplace op followed by an overwrite expression and then in a later pass replacing the overwrite expression with a inplace copy or just updating the subsequent uses of the overwritten mlir::Value (this is implemented in `createRemoveOverwritePass()`).

For view operations things are a little more complicated. The view operations are added to the graph as normal (although if the operation is inplace the builder is updated immediately with the new mlir::Value - this is because the new value could potentially have a different shape). A second pass then converts each use of the view into an outplace view operation and each overwrite (i.e. when an inplace op has been applied to the view) with a inverse view operation (i.e. an op that recovers the original tensor from the view) and an overwrite of the original tensor. This pass is applied recursively until every operation on a view has been outplaced. After this the original reference view operation will be eliminated by dead code elimination. This operation will add more outplace views than necessary but these are be removed by CSE. This pass is done in `createOutplaceViewOpsPass()`.

Note that the view operations pass relies on having the overwrite operations in the graph so the passes which remove these must occur after the pass that outplaces the view operations. To facilitate adding these in the correct order `addOverwriteHandlingPasses()` has been created.

Test Plan:
* mlir_view_operations.py has been created to test the view operations
* mlir_dce.py has been created to test the dead code elimination

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, nicholasw, ahmads, georgew

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, nicholasw, ahmads, georgew

Subscribers: georgew, ahmads, charliew, nicholasw, anthonyb

Maniphest Tasks: T62743, T65711, T65602

Differential Revision: https://phabricator.sourcevertex.net/D74708