Set the default values of options from environment
Summary:
Use a POPTORCH_DEFAULT_OPTIONS variable to control the default
value of any poptorch.Options object.
This essentially implements JSON parsing of the Options and passes
in the string from the environment variable.

Motivation
In CI we want our tutorials to run with OnDemand IPUs to improve utilisation
but we don't necessarily want users to run in this mode.
The abiliy to set default values for options from an environment variable
will let us set our CI to on demand IPU acquisition
and improve our machine utilisation.

Fixes T69998

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, davidw

Maniphest Tasks: T69998

Differential Revision: https://phabricator.sourcevertex.net/D75850