Infrastructure for MLIR pass testing
Summary:
Contains an `mlir-opt` equivalent, cmake hook-ups so that the tests run with the full suite, and some basic cases to test the tests.

The tests that are in this diff are mostly a demonstration/for testing things out, and aren't exhaustive -- I'm going to make another diff that fills them out.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, georgew

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, georgew

Maniphest Tasks: T69216

Differential Revision: https://phabricator.sourcevertex.net/D76616