Refactored eager tensor store out of the eager compiler
Summary: Instead of storing the tensor data for eager mode in the compiler allocate them in the TensorStore instead (in the same way that the static compiler works). This means that we can reset the compiler between each run which makes it easier to do funciton caching.

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, georgew, anthonyb

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, georgew, anthonyb

Subscribers: anthonyb, georgew

Maniphest Tasks: T67886, T64272, T66925

Differential Revision: https://phabricator.sourcevertex.net/D76478